python/requirements.txt
certifi==2019.11.28
chardet==3.0.4
dbus-python==1.2.16
idna==2.8
influxdb-client==1.35.0
minimalmodbus==2.0.1
paho-mqtt==1.6.1
pycairo==1.16.2
PyGObject==3.36.0
pyserial==3.5
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
reactivex==4.0.4
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
typing-extensions==4.4.0
urllib3==1.26.13
